requirements.txt
# Core web framework
Flask

# HTTP requests for fetching TLEs
requests

# Astronomy & orbital mechanics
skyfield
sgp4

# Timezone support (zoneinfo is built-in for Python >=3.9)
# Only include this if running on Python < 3.9
# zoneinfo; python_version < "3.9"

# Optional: if you later add automatic timezone detection
timezonefinder==6.2.0
psutil
schedule
python-dateutil
Pillow==10.2.0
scipy
numpy
pydub


# SSTV decoder from GitHub

git+https://github.com/colaclanth/sstv.git

# Note: PD120-specific decoders were removed due to integration issues.

## Optional alternative decoders (commented out)
# git+https://github.com/colaclanth/rxsstv.git  # requires sign in
